src/test/javascript/spec/app/entities/frog/frog-update.component.spec.ts
/* tslint:disable max-line-length */
import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils';
import sinon, { SinonStubbedInstance } from 'sinon';
import Router from 'vue-router';
import { ToastPlugin } from 'bootstrap-vue';

import * as config from '@/shared/config/config';
import FrogUpdateComponent from '@/entities/frog/frog-update.vue';
import FrogClass from '@/entities/frog/frog-update.component';
import FrogService from '@/entities/frog/frog.service';

import AlertService from '@/shared/alert/alert.service';

const localVue = createLocalVue();

config.initVueApp(localVue);
const i18n = config.initI18N(localVue);
const store = config.initVueXStore(localVue);
const router = new Router();
localVue.use(Router);
localVue.use(ToastPlugin);
localVue.component('font-awesome-icon', {});
localVue.component('b-input-group', {});
localVue.component('b-input-group-prepend', {});
localVue.component('b-form-datepicker', {});
localVue.component('b-form-input', {});

describe('Component Tests', () => {
  describe('Frog Management Update Component', () => {
    let wrapper: Wrapper<FrogClass>;
    let comp: FrogClass;
    let frogServiceStub: SinonStubbedInstance<FrogService>;

    beforeEach(() => {
      frogServiceStub = sinon.createStubInstance<FrogService>(FrogService);

      wrapper = shallowMount<FrogClass>(FrogUpdateComponent, {
        store,
        i18n,
        localVue,
        router,
        provide: {
          frogService: () => frogServiceStub,
          alertService: () => new AlertService(),
        },
      });
      comp = wrapper.vm;
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const entity = { id: 123 };
        comp.frog = entity;
        frogServiceStub.update.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(frogServiceStub.update.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        comp.frog = entity;
        frogServiceStub.create.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(frogServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });

    describe('Before route enter', () => {
      it('Should retrieve data', async () => {
        // GIVEN
        const foundFrog = { id: 123 };
        frogServiceStub.find.resolves(foundFrog);
        frogServiceStub.retrieve.resolves([foundFrog]);

        // WHEN
        comp.beforeRouteEnter({ params: { frogId: 123 } }, null, cb => cb(comp));
        await comp.$nextTick();

        // THEN
        expect(comp.frog).toBe(foundFrog);
      });
    });

    describe('Previous state', () => {
      it('Should go previous state', async () => {
        comp.previousState();
        await comp.$nextTick();

        expect(comp.$router.currentRoute.fullPath).toContain('/');
      });
    });
  });
});
